+++ content/js/dmelearn.js
@@ -99,7 +99,7 @@ $(function() {
             window.location.href = link;
         });
     });
-    // Collapsable elements.
+    // Collapsible elements.
     // Swap arrow sprite on hide/show.
     $('.accordion').on('show hide', function(n) {
         $(n.target).siblings('.accordion-heading').find('.accordion-toggle i').toggleClass('icon-chevron-up icon-chevron-down');