+++ content/plates/layouts/3/footer.php
@@ -1,0 +1,8 @@
+<?php //Course Footer ?>
+<div class="footer">
+</div>
+<?php //Jquery-UI ?>
+<script src="<?=$constants['base_js']?>jquery-ui.min.js"></script>
+<script src="<?=$constants['base_js']?>jquery.ui.touch-punch.min.js"></script>
+<?php //Bootstrap 3 JS  ?>
+<script src="js/bootstrap.min.js"></script>