+++ content/plates/layouts/3/shell.php
@@ -1,0 +1,4 @@
+<?php //Course shell version 3 ?>
+<div class="container">
+    <br><p>This course is not supported in this version of the dmelearn plugin. Please contact your Moodle administrator for more information.</p>
+</div>