+++ content/plates/partials/html.php
@@ -1,4 +1,3 @@
-{# <html>’s conditional classes #}
 <!--[if lte IE 7]><html class="no-js lt-ie10 lt-ie9 lt-ie8 lt-ie7 lte-ie7 mdl-dmelearn" lang="en"><![endif]-->
 <!--[if IE 8]><html class="no-js lt-ie10 lt-ie9 ie8 mdl-dmelearn" lang="en"><![endif]-->
 <!--[if IE 9]><html class="no-js lt-ie10 ie9 mdl-dmelearn" lang="en"><![endif]-->